requirements.txt
requests
bs4
pypyodbc
urllib3
pandas
pythainlp
boto3
google-cloud-vision